Using Artificial Neural Networks for flower recognition
This project aims was to build an artificial neural network for flower recognition. The neural network was trained on a dataset of flowers and the performances of the techniques used was compared. The techniques used were CNN and ResNet50.

Technologies used: Python, Jupyter Notebook, CNN, Transfer Learning, keras, numpy, pandas, sklearn, tensorflow, fastai, matplotlib.